♻️ extract useUnpaidIssues
